+++ tests/testthat/test-translate_sql.R
@@ -425,3 +425,22 @@ with_locale(test.locale(), test_that)("first(), last(), and nth() work", {
     dbplyr::sql('NTH_VALUE("x", 2) IGNORE NULLS OVER (ORDER BY "y")')
   )
 })
+
+with_locale(test.locale(), test_that)("paste() works", {
+  s <- setup_mock_dplyr_connection()[["db"]]
+
+  expect_equal(
+    dbplyr::translate_sql(paste(a, b, c), con = s$con),
+    dbplyr::sql("\"a\" || ' ' || \"b\" || ' ' || \"c\"")
+  )
+
+  expect_equal(
+    dbplyr::translate_sql(paste(a, b, c, sep = "-"), con = s$con),
+    dbplyr::sql("\"a\" || '-' || \"b\" || '-' || \"c\"")
+  )
+
+  expect_equal(
+    dbplyr::translate_sql(paste0(a, b, c), con = s$con),
+    dbplyr::sql("\"a\" || \"b\" || \"c\"")
+  )
+})